START View fourth-app/programming_blog/

+++ third-app/todos/todo/views.py
@@ -6,13 +6,13 @@ from django.contrib.auth import login, logout, authenticate
 from django.utils import timezone
 from django.contrib.auth.decorators import login_required
 
-
 from .forms import TodoForm
 from .models import Todo
 
 
 def home(request):
     return render(request, 'todo/home.html')
+
 
 def signupuser(request):
     if request.method == 'GET':
@@ -21,7 +21,8 @@ def signupuser(request):
     else:
         if request.POST['password1'] == request.POST['password2']:
             try:
-                user = User.objects.create_user(request.POST['username'], password=request.POST['password1'])
+                user = User.objects.create_user(
+                    request.POST['username'], password=request.POST['password1'])
                 user.save()
                 login(request, user)
                 return redirect('currenttodos')
@@ -29,10 +30,11 @@ def signupuser(request):
             except IntegrityError:
                 return render(request, 'todo/signupuser.html',
                               {'form': UserCreationForm(), 'error': 'Такое имя '
-                                                'пользователя уже существует. Задайте Другое'})
+                                                                    'пользователя уже существует. Задайте Другое'})
         else:
             return render(request, 'todo/signupuser.html',
                           {'form': UserCreationForm(), 'error': 'Пароли не Совпадают.'})
+
 
 def logoutuser(request):
     if request.method == 'POST':
@@ -40,9 +42,11 @@ def logoutuser(request):
         return redirect('home')
 
 
+@login_required()
 def currenttodos(request):
     todos = Todo.objects.filter(user=request.user, date_completed__isnull=True)
     return render(request, 'todo/currenttodos.html', {'todos': todos})
+
 
 def loginuser(request):
     if request.method == 'GET':
@@ -52,14 +56,15 @@ def loginuser(request):
                             password=request.POST['password'])
         if user is None:
             return render(request, 'todo/loginuser.html', {'form': AuthenticationForm(),
-                                                            'error': 'Неверные данные'})
+                                                           'error': 'Неверные данные'})
         else:
             login(request, user)
             return redirect('currenttodos')
 
 
+@login_required()
 def createtodo(request):
-    if request.method =='GET':
+    if request.method == 'GET':
         return render(request, 'todo/createtodo.html', {'form': TodoForm()})
     else:
         try:
@@ -67,17 +72,20 @@ def createtodo(request):
             new_todo = form.save(commit=False)
             new_todo.user = request.user
             new_todo.save()
-            return redirect("curenttodos")
+            return redirect("currenttodos")
         except ValueError:
-            return redirect(request, 'todo/createtodo.html', {'form': TodoForm(),
-                                                              'error': 'Переданы неверные данныеюПопр еще раз'})
+            return render(request, 'todo/createtodo.html',
+                          {'form': TodoForm(),
+                           'error': 'Переданы неверные данныеюПопр еще раз'})
 
+
+@login_required()
 def viewtodo(request, todo_pk):
     todo = get_object_or_404(Todo, pk=todo_pk)
     if request.method == 'GET':
         form = TodoForm(instance=todo)
         return render(request, 'todo/viewtodo.html', {'todo': todo,
-                                                  'form': form})
+                                                      'form': form})
     else:
         try:
             form = TodoForm(request.POST, instance=todo)
@@ -85,8 +93,8 @@ def viewtodo(request, todo_pk):
             return redirect('currenttodos')
         except ValueError:
             return render(request, 'todo/viewtodo.html', {'todo': todo,
+                                                          'form': form, 'error': 'Неверные данные'})
 
-                                             'form': form, 'erreo': 'Not correct Data'})
 
 @login_required
 def completetodo(request, todo_pk):
@@ -96,18 +104,18 @@ def completetodo(request, todo_pk):
         todo.save()
         return redirect('currenttodos')
 
-@ login_required
+
+@login_required
 def deletetodo(request, todo_pk):
     todo = get_object_or_404(Todo, pk=todo_pk, user=request.user)
-    todo.date_completed = timezone.now()
-    todo.delete()
-    return redirect('currenttodos')
+    if request.method == 'POST':
+        todo.delete()
+        return redirect('currenttodos')
+
 
 @login_required
 def completedtodo(request):
-    todos = Todo.objects.\
-        filter(user=request.user, date_completed__isnull=False).\
+    todos = Todo.objects. \
+        filter(user=request.user, date_completed__isnull=False). \
         order_by('-date_completed')
-    return render(request, 'todo/completedtodo')
-
-
+    return render(request, 'todo/completedtodo.html', {'todos': todos})
